+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 ### Changed
 
 - The location of the file `spotdll.json` is now called from the folder the script is being called from, rather than from `~/bin`. #1
+- Added a switch to generate the `spotdll.json` file from sample in the Current working directory.
 
 ### Issues Closed
 
 - #1
+- #3 WIP
 
 ---